pyproject.toml
[project]
name = "greet-fact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "duckdb>=1.3.1",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.16",
    "langchain-google-genai>=2.1.6",
    "langchain-groq>=0.3.4",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "setuptools>=80.9.0",
]
